tweaked

+++ DystopiaEngine/DystopiaEngine/Source/Editor/Dock.cpp
@@ -1360,7 +1360,7 @@ static void readLine(ImGuiContext* /*ctx*/, ImGuiSettingsHandler* /*handler*/, v
 		int prev, next, child0, child1, mpParentTab;
 		char label[64], mLocation[64];
 
-		if (sscanf_s(line_start, "label=%[^\n^\r]", label, _countof(label)) == 1)
+		if (sscanf_s(line_start, "label=%[^\n^\r]", label, static_cast<unsigned int>(_countof(label))) == 1)
 		{
 			userdata->dock->mpLabel = ImStrdup(label);
 			userdata->dock->mId = ImHash(userdata->dock->mpLabel, 0);
@@ -1389,7 +1389,7 @@ static void readLine(ImGuiContext* /*ctx*/, ImGuiSettingsHandler* /*handler*/, v
 		{
 			userdata->dock->mOpened = (opened) ? true : false ;
 		}
-		else if (sscanf_s(line_start, "mLocation=%[^\n^\r]", mLocation, _countof(mLocation)) == 1)
+		else if (sscanf_s(line_start, "mLocation=%[^\n^\r]", mLocation, static_cast<unsigned int>(_countof(mLocation))) == 1)
 		{
 			strcpy_s(userdata->dock->mLocation, mLocation);
 		}
